Change clunky loop, use genexpr instead
Using a loop like this to construct a new value is generally bad.

It used to be common in python way back, but since everyone used it all the time
we got a new (since python 2.0, which came out on October 16, 2000) way of doing
this: list comprehensions / generator expressions.

So I refactored it to that. I hope it is easier to understand and more pleasing
to the eye.

Also generators / list comprehensions can be better than loops because
they should be faster (at least that is what most python programmers
claim).

Generators can be better, because they support some lazy evaluation.
(I.E you can choose to delay evaluating your code)