Further fixes
The model starts now, but won't run yet (instant crash).

However when logging output in portrayCell we sometimes get a cell.state as int
and sometimes we get an enum?

Also some cells are still white, which should not be the case, because the white
coloring means that it did not get assigned a CellState.

Which means that we have missed fixing a cell.state assignment somewhere.