Fix std mean
Apparently pythons mean function cannot handle iterators.
The std python satistics.mean will run out the iterator and then throw an
exception, when it is empty.

A nice side effect is that this will terminate the simulation when there are no
more infected cells. (Which is an end state).

However I supplied my own function.